+++ packages/common/src/utils/hooks/onUpdate.ts
@@ -1,8 +1,0 @@
-import { useEffect } from 'react'
-import { AnyFunction } from '../../types'
-
-export const onUpdate = (func: AnyFunction, listeners = []) => {
-  useEffect(() => {
-    return func()
-  }, listeners)
-}