+++ packages/common/src/utils/hooks/useForceRender.ts
@@ -1,6 +1,0 @@
-import { useReducer } from 'react'
-
-export function useForceRender() {
-  const [_, forceRender] = useReducer((x) => x + 1, 0)
-  return forceRender
-}