+++ packages/hooks/src/index.ts
@@ -33,6 +33,7 @@ export * from './useSearch'
 export * from './usePartialState'
 export * from './useIsClient'
 export * from './GlobalContext'
+export * from './useComponentTestId'
 
 export {
   useEffect,