+++ packages/mobile/src/components/Navigation/index.tsx
@@ -1,3 +1,0 @@
-export * from './Navigation'
-export * from './types'
-export * from './utils'