+++ packages/mobile/src/hooks/index.ts
@@ -1,13 +1,3 @@
-import { StyleProp, useStyleObserver } from '@codeleap/styles'
-import { useMemo } from 'react'
-import { MobileStyleRegistry } from '../Registry'
-
-export const useStylesFor = <T = unknown>(componentName: string, style: StyleProp<T, string>): T => {
-  const styleObserver = useStyleObserver(style)
-
-  const styles = useMemo(() => {
-    return MobileStyleRegistry.current.styleFor(componentName, style)
-  }, [styleObserver])
-
-  return styles
-}
+export * from './useStylesFor'
+export * from './useStatusBar'
+export * from './useKeyboardController'