+++ packages/mobile/src/utils/KeyboardAware/index.ts
@@ -1,4 +1,4 @@
 
 export * from './keyboardHooks'
 export * from './context'
-export * from './types'
+