+++ syncpack.config.js
@@ -13,19 +13,19 @@ module.exports = {
       "label": "React",
       "dependencies": ["react"],
       packages: ["**"],
-      pinVersion: "18.1.0"
+      pinVersion: "18.2.0"
     },
     {
       "label": "Typescript",
       "dependencies": ["typescript"],
       packages: ["**"],
-      pinVersion: "5.0.4"
+      pinVersion: "5.5.2"
     },
     {
       "label": "React native",
       "dependencies": ["react-native"],
       packages: ["**"],
-      pinVersion: "0.76.1"
+      pinVersion: "0.77.0"
 
     },
     {