+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [4.3.9](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.6...v4.3.9) (2025-01-22)
+
+
+### Bug Fixes
+
+* **styles:** createStyles types ([0ed5ff8](https://github.com/codeleap-uk/internal-libs-monorepo/commit/0ed5ff828882be459dffd05d6d0646429fb82234))
+* **styles:** undefined theme ([1fa62d7](https://github.com/codeleap-uk/internal-libs-monorepo/commit/1fa62d76c79d59147b1cf6a12c4d77be529dfd87))
+
+### [4.3.8](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.7...v4.3.8) (2025-01-21)
+
+
+### Bug Fixes
+
+* **styles:** createStyles types ([0ed5ff8](https://github.com/codeleap-uk/internal-libs-monorepo/commit/0ed5ff828882be459dffd05d6d0646429fb82234))
+
+### [4.3.7](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.5...v4.3.7) (2025-01-21)
+
+
+### Bug Fixes
+
+* conditional state check ([2b3c7de](https://github.com/codeleap-uk/internal-libs-monorepo/commit/2b3c7deacfc2802b1591925d9ff3573b449f2920))
+* **styles:** undefined theme ([1fa62d7](https://github.com/codeleap-uk/internal-libs-monorepo/commit/1fa62d76c79d59147b1cf6a12c4d77be529dfd87))
+
+### [4.3.6](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.2.18...v4.3.6) (2025-01-01)
+
+
+### Bug Fixes
+
+* conditional state check ([2b3c7de](https://github.com/codeleap-uk/internal-libs-monorepo/commit/2b3c7deacfc2802b1591925d9ff3573b449f2920))
+* **styles:** web border ([b0e29d6](https://github.com/codeleap-uk/internal-libs-monorepo/commit/b0e29d6f40e8b0f528d9f12eb6b4886a38940481))
+
+### [4.3.5](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.3...v4.3.5) (2024-12-31)
+
+
+### Bug Fixes
+
+* **styles:** web border ([b0e29d6](https://github.com/codeleap-uk/internal-libs-monorepo/commit/b0e29d6f40e8b0f528d9f12eb6b4886a38940481))
+
+### [4.3.4](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.3...v4.3.4) (2024-12-31)
+
+
+### Bug Fixes
+
+* **styles:** web border ([b0e29d6](https://github.com/codeleap-uk/internal-libs-monorepo/commit/b0e29d6f40e8b0f528d9f12eb6b4886a38940481))
+
+### [4.3.3](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.2...v4.3.3) (2024-12-16)
+
 ### [4.3.2](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.1...v4.3.2) (2024-12-12)
 
 ### [4.3.1](https://github.com/codeleap-uk/internal-libs-monorepo/compare/v4.3.0...v4.3.1) (2024-12-09)